requirements.txt
torch>=0.4
torchvision
tensorboardX
pillow